fix client #7
